pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testgen"
version = "1.0.0"
description = "AI-powered test case generation system with clean architecture"
authors = [{ name = "HieuDev" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["ai", "testing", "test-generation", "clean-architecture"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Testing",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# --- RUNTIME deps ---
dependencies = [
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.20.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "openai>=1.0.0",
  "python-json-logger>=2.0.7",
  "click>=8.0.0",
  "python-multipart>=0.0.20",
  "markitdown[all]>=0.1.2",
  "openpyxl>=3.1.2",
  "rootutils>=1.0.7",
  "networkx>=3.0",
  "pandas>=2.3.1",
  "omegaconf>=2.3.0",
  "rank-bm25>=0.2.2",
  "numpy>=1.26",
  "scikit-learn>=1.3.0",
  "blake3>=1.0.5",
  "pre-commit>=4.3.0",
  "sqlalchemy[asyncio]>=2.0.43",
  "alembic>=1.16.5",
  "psycopg[binary]>=3.2.10",
  "asyncpg>=0.30.0",
  "google-adk>=1.14.0",
  "python-dotenv>=1.1.1",
  "google-generativeai>=0.8.0",
  "pgvector>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.7",
  "mypy>=1.5.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.11.0",
  "httpx>=0.24.0",
  "ruff>=0.5.0",
]
notebook = ["ipykernel>=6.30.1"]
viz = ["matplotlib>=3.8"]
yaml = ["PyYAML>=6.0.0"]
monitoring = ["prometheus-client>=0.17.0"]

all = ["testgen[dev,notebook,viz,yaml,monitoring]"]

[project.urls]
Homepage = "https://github.com/ai-testgen/enterprise"
Documentation = "https://docs.aitestgen.com"
Repository = "https://github.com/ai-testgen/enterprise.git"
Issues = "https://github.com/ai-testgen/enterprise/issues"

[project.scripts]
ai-testgen = "app.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "testgen*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools]
include-package-data = true

# ---------------- Tooling ----------------
[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["migrations", "build", "dist", ".venv", "venv", "logs", "data"]

# Schema mới của Ruff: không dùng [tool.ruff.lint] nữa
select = ["E","F","I","UP","B","SIM","C4","PIE"]
ignore = ["D"]

[tool.ruff.isort]
known-first-party = ["app","src","testgen"]
combine-as-imports = true
force-wrap-aliases = true

[tool.mypy]
python_version = "3.12"
pretty = true
show_error_codes = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
ignore_missing_imports = true
exclude = "(^migrations/|^tests?/fixtures/|^venv/|^.venv/)"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q -ra"
testpaths = ["tests"]
xfail_strict = true
